+++ src/lib/shared.ts
@@ -1,5 +1,3 @@
-import { error } from "itty-router-extras";
-
 export const hexKeyRegex4Digits = /^0x[a-f0-9]{4}$/;
 export const hexKeyRegex2Digits = /^0x[a-f0-9]{2}$/;
 export const firmwareVersionRegex = /^\d{1,3}\.\d{1,3}$/;
@@ -49,6 +47,13 @@ export function padVersion(version: string): string {
 	return version + ".0";
 }
 
+// expands object types recursively
+export type ExpandRecursively<T> = T extends object
+	? T extends infer O
+		? { [K in keyof O]: ExpandRecursively<O[K]> }
+		: never
+	: T;
+
 export function array2hex(arr: Uint8Array): string {
 	return [...arr].map((x) => x.toString(16).padStart(2, "0")).join("");
 }
@@ -61,50 +66,3 @@ export function hex2array(hex: string): Uint8Array {
 	}
 	return ret;
 }
-
-/** Constant-time string comparison */
-export function safeCompare(expected: string, actual: string): boolean {
-	const lenExpected = expected.length;
-	let result = 0;
-
-	if (lenExpected !== actual.length) {
-		actual = expected;
-		result = 1;
-	}
-
-	for (let i = 0; i < lenExpected; i++) {
-		result |= expected.charCodeAt(i) ^ actual.charCodeAt(i);
-	}
-
-	return result === 0;
-}
-
-// expands object types recursively
-export type ExpandRecursively<T> = T extends object
-	? T extends infer O
-		? { [K in keyof O]: ExpandRecursively<O[K]> }
-		: never
-	: T;
-
-export function clientError(
-	message: BodyInit | Record<string, any> | undefined,
-	code: number = 400
-): Response {
-	return error(code, message);
-}
-
-export function serverError(
-	message: BodyInit | Record<string, any>,
-	code: number = 500
-): Response {
-	return error(code, message);
-}
-
-export type RequestWithProps<
-	U extends Record<string, unknown>[],
-	T extends Request = Request
-> = U extends [infer First, ...infer Rest extends Record<string, unknown>[]]
-	? RequestWithProps<Rest, T & First>
-	: T;
-
-export type ContentProps = { content: unknown };
